+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Rule Engine classes not found, because the jars is inside a container jar, we need to extract the file "ch.ivyteam.ivy.rule.engine.libs" manually. Only for Axon.ivy Engine 7.0.
 - Index of out bound because we assume all modules have at least one content root. It turned out some Gradle modules don't have it.
 - Adjusted the pattern to extract Axon.ivy Engine port from console log. The current pattern will work correctly on Axon.ivy Engine 7.0 and 8.0.
+- Remaining half number of the global variables / system properties, cause them duplicated in the EngineView.
 
 ## [v0.2.1] - 2021-02-03
 ### Changed